refactor(db): rename Store schema to Room (SA2-42)
- store テーブル/カラムを room へリネーム（schema/room.ts に分離）
- schema/store.ts に同居していた currency 系3テーブルを schema/currency.ts へ分離
- ring_game / tournament / game_session の store_id → room_id（FK・index も改名）
- 0029_rename_store_to_room.sql を追加（pure ALTER ... RENAME / データ保持）
- 関連テストを room へ更新（local D1 へのマイグレーション適用を確認済み）
