+++ packages/db/src/schema.ts
@@ -19,7 +19,16 @@ import {
 	playerToPlayerTag,
 	playerToPlayerTagRelations,
 } from "./schema/player";
+import {
+	currency,
+	currencyRelations,
+	currencyTransaction,
+	currencyTransactionRelations,
+	transactionType,
+	transactionTypeRelations,
+} from "./schema/currency";
 import { ringGame, ringGameRelations } from "./schema/ring-game";
+import { room, roomRelations } from "./schema/room";
 import { gameSession, gameSessionRelations } from "./schema/session";
 import {
 	sessionBlindLevel,
@@ -49,16 +58,6 @@ import {
 	sessionTournamentDetailRelations,
 } from "./schema/session-tournament-detail";
 import {
-	currency,
-	currencyRelations,
-	currencyTransaction,
-	currencyTransactionRelations,
-	store,
-	storeRelations,
-	transactionType,
-	transactionTypeRelations,
-} from "./schema/store";
-import {
 	blindLevel,
 	blindLevelRelations,
 	tournament,
@@ -80,8 +79,8 @@ export const schema = {
 	user,
 	userRelations,
 	verification,
-	store,
-	storeRelations,
+	room,
+	roomRelations,
 	currency,
 	currencyRelations,
 	transactionType,
