+++ packages/db/src/schema/room.ts
@@ -1,0 +1,29 @@
+import { relations, sql } from "drizzle-orm";
+import { index, integer, sqliteTable, text } from "drizzle-orm/sqlite-core";
+import { user } from "./auth";
+
+export const room = sqliteTable(
+	"room",
+	{
+		id: text("id").primaryKey(),
+		userId: text("user_id")
+			.notNull()
+			.references(() => user.id, { onDelete: "cascade" }),
+		name: text("name").notNull(),
+		memo: text("memo"),
+		createdAt: integer("created_at", { mode: "timestamp" })
+			.default(sql`(unixepoch())`)
+			.notNull(),
+		updatedAt: integer("updated_at", { mode: "timestamp" })
+			.$onUpdate(() => /* @__PURE__ */ new Date())
+			.notNull(),
+	},
+	(table) => [index("room_userId_idx").on(table.userId)]
+);
+
+export const roomRelations = relations(room, ({ one }) => ({
+	user: one(user, {
+		fields: [room.userId],
+		references: [user.id],
+	}),
+}));
